move if-statement interpreter tests to separate file
