+++ runtime/tests/interpreter/interpreter_test.go
@@ -1154,74 +1154,6 @@ func TestInterpretAndOperatorShortCircuitLeftFailure(t *testing.T) {
 	)
 }
 
-func TestInterpretIfStatement(t *testing.T) {
-
-	t.Parallel()
-
-	inter := parseCheckAndInterpretWithOptions(t,
-		`
-           pub fun testTrue(): Int {
-               if true {
-                   return 2
-               } else {
-                   return 3
-               }
-               return 4
-           }
-
-           pub fun testFalse(): Int {
-               if false {
-                   return 2
-               } else {
-                   return 3
-               }
-               return 4
-           }
-
-           pub fun testNoElse(): Int {
-               if true {
-                   return 2
-               }
-               return 3
-           }
-
-           pub fun testElseIf(): Int {
-               if false {
-                   return 2
-               } else if true {
-                   return 3
-               }
-               return 4
-           }
-        `,
-		ParseCheckAndInterpretOptions{
-			HandleCheckerError: func(err error) {
-				errs := checker.ExpectCheckerErrors(t, err, 2)
-
-				assert.IsType(t, &sema.UnreachableStatementError{}, errs[0])
-				assert.IsType(t, &sema.UnreachableStatementError{}, errs[1])
-			},
-		},
-	)
-
-	for name, expected := range map[string]int64{
-		"testTrue":   2,
-		"testFalse":  3,
-		"testNoElse": 2,
-		"testElseIf": 3,
-	} {
-		t.Run(name, func(t *testing.T) {
-			value, err := inter.Invoke(name)
-			require.NoError(t, err)
-
-			assert.Equal(t,
-				interpreter.NewIntValueFromInt64(expected),
-				value,
-			)
-		})
-	}
-}
-
 func TestInterpretExpressionStatement(t *testing.T) {
 
 	t.Parallel()
@@ -3328,208 +3260,6 @@ func TestInterpretCompositeNilEquality(t *testing.T) {
 			)
 		})
 	}
-}
-
-func TestInterpretIfStatementTestWithDeclaration(t *testing.T) {
-
-	t.Parallel()
-
-	inter := parseCheckAndInterpret(t, `
-      var branch = 0
-
-      fun test(x: Int?): Int {
-          if var y = x {
-              branch = 1
-              return y
-          } else {
-              branch = 2
-              return 0
-          }
-      }
-    `)
-
-	t.Run("2", func(t *testing.T) {
-		value, err := inter.Invoke(
-			"test",
-			interpreter.NewIntValueFromInt64(2),
-		)
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(2),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(1),
-			inter.Globals["branch"].Value,
-		)
-	})
-
-	t.Run("nil", func(t *testing.T) {
-		value, err := inter.Invoke("test", interpreter.NilValue{})
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(0),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(2),
-			inter.Globals["branch"].Value,
-		)
-	})
-}
-
-func TestInterpretIfStatementTestWithDeclarationAndElse(t *testing.T) {
-
-	t.Parallel()
-
-	inter := parseCheckAndInterpret(t, `
-      var branch = 0
-
-      fun test(x: Int?): Int {
-          if var y = x {
-              branch = 1
-              return y
-          }
-          branch = 2
-          return 0
-      }
-    `)
-
-	t.Run("2", func(t *testing.T) {
-		value, err := inter.Invoke(
-			"test",
-			interpreter.NewIntValueFromInt64(2),
-		)
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(2),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(1),
-			inter.Globals["branch"].Value,
-		)
-	})
-
-	t.Run("nil", func(t *testing.T) {
-		value, err := inter.Invoke("test", interpreter.NilValue{})
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(0),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(2),
-			inter.Globals["branch"].Value,
-		)
-
-	})
-}
-
-func TestInterpretIfStatementTestWithDeclarationNestedOptionals(t *testing.T) {
-
-	t.Parallel()
-
-	inter := parseCheckAndInterpret(t, `
-      var branch = 0
-
-      fun test(x: Int??): Int? {
-          if var y = x {
-              branch = 1
-              return y
-          } else {
-              branch = 2
-              return 0
-          }
-      }
-    `)
-
-	t.Run("2", func(t *testing.T) {
-		value, err := inter.Invoke(
-			"test",
-			interpreter.NewIntValueFromInt64(2),
-		)
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewSomeValueOwningNonCopying(
-				interpreter.NewIntValueFromInt64(2),
-			),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(1),
-			inter.Globals["branch"].Value,
-		)
-	})
-
-	t.Run("nil", func(t *testing.T) {
-		value, err := inter.Invoke("test", interpreter.NilValue{})
-		require.NoError(t, err)
-
-		assert.Equal(t,
-			interpreter.NewSomeValueOwningNonCopying(
-				interpreter.NewIntValueFromInt64(0),
-			),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(2),
-			inter.Globals["branch"].Value,
-		)
-	})
-}
-
-func TestInterpretIfStatementTestWithDeclarationNestedOptionalsExplicitAnnotation(t *testing.T) {
-
-	t.Parallel()
-
-	inter := parseCheckAndInterpret(t, `
-      var branch = 0
-
-      fun test(x: Int??): Int? {
-          if var y: Int? = x {
-              branch = 1
-              return y
-          } else {
-              branch = 2
-              return 0
-          }
-      }
-    `)
-
-	t.Run("2", func(t *testing.T) {
-		value, err := inter.Invoke(
-			"test",
-			interpreter.NewIntValueFromInt64(2),
-		)
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewSomeValueOwningNonCopying(
-				interpreter.NewIntValueFromInt64(2),
-			),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(1),
-			inter.Globals["branch"].Value,
-		)
-
-	})
-
-	t.Run("nil", func(t *testing.T) {
-		value, err := inter.Invoke("test", interpreter.NilValue{})
-		require.NoError(t, err)
-		assert.Equal(t,
-			interpreter.NewSomeValueOwningNonCopying(
-				interpreter.NewIntValueFromInt64(0),
-			),
-			value,
-		)
-		assert.Equal(t,
-			interpreter.NewIntValueFromInt64(2),
-			inter.Globals["branch"].Value,
-		)
-	})
 }
 
 func TestInterpretInterfaceConformanceNoRequirements(t *testing.T) {
